refactored chainfilter.rs, split out indexer
